Move date util functions
