+++ corehq/apps/accounting/forms.py
@@ -64,7 +64,6 @@ from corehq.apps.accounting.models import (
 )
 from corehq.apps.accounting.tasks import send_subscription_reminder_emails
 from corehq.apps.accounting.utils import (
-    get_first_last_days,
     get_money_str,
     has_subscription_already_ended,
     make_anchor_tag,
@@ -73,6 +72,7 @@ from corehq.apps.domain.models import Domain
 from corehq.apps.hqwebapp.crispy import BootstrapMultiField, TextField
 from corehq.apps.hqwebapp.tasks import send_html_email_async
 from corehq.apps.users.models import WebUser
+from corehq.util.dates import get_first_last_days
 
 
 class BillingAccountBasicForm(forms.Form):
