+++ corehq/apps/accounting/tests/test_invoice_factory.py
@@ -8,7 +8,7 @@ from corehq.apps.accounting.models import (
     SubscriptionAdjustment,
 )
 from corehq.apps.accounting.tests.base_tests import BaseAccountingTest
-from corehq.apps.accounting.utils import get_previous_month_date_range
+from corehq.util.dates import get_previous_month_date_range
 
 
 class TestDomainInvoiceFactory(BaseAccountingTest):
