+++ corehq/apps/accounting/tests/test_models.py
@@ -22,7 +22,7 @@ from corehq.apps.smsbillables.models import (
     SmsBillable,
 )
 from corehq.apps.accounting.tests.base_tests import BaseAccountingTest
-from corehq.apps.accounting.utils import get_previous_month_date_range
+from corehq.util.dates import get_previous_month_date_range
 
 
 class TestBillingAccount(BaseAccountingTest):
